Implemented TestAddItem Class to test Bill class
Implemented several test methods which test the constraints and
adherence of the addItem method.

Note - A problem has occured in my file explorer where mutliple lab 2
directories exist and I have trouble using commit on the right
directory. The previous commit may have been on the false directory and
thus this commit seems separate and not a continuation.

This is commit #2, following the commit with the intial creation of test
classes