3B Completed: Task.java updated with desc setter
setDesc() has been added to Task.java to update the description of a
task. I recognize the discrepancy between getDescription() and
setDesc(), with setDesc using Desc and the getter not using Desc. I
thought about changing the getter to getDesc() but decided against it
because I didn't feel like updating it everywhere else. Sorry