Task 1A completed: Persistent Sort functionality
Added code to check the current selected sorting order from the combobox
and resort the task list every time a task is added. The 1a task prompt
says to also maintain sort when a task is "updated in any way" but I am
not sure what is meant by that. Updating a description shouldn't alter
sort but just in case I also modified the updateDescription method.