Project 3 Subtask 3B complete
I struggled with this one. Where do I put the map...? in a model file or
the viewmodel file...? Both ways involved O(n) time with n being number
of collections but I decided I would put the map in the viewmodel just
for simplicity. That left me with pretty much no changes to be made to
model files so I just added a finder method to Collection.java if I need
to return a comic object after verifying it exists with the map search
(I probably won't use this since I will return strings in the pop-up
window, no need to find the actual comic object if we have its name and
issue from the map already)